requirements.txt
python-dotenv==0.20.0
kafka-python==2.0.2
faker==18.9.0